actualizacion, no existe double, se probara con el campo float

+++ new_module/models/models.py
@@ -78,7 +78,7 @@ class pruebas(models.Model):
     )
     montoNeto = fields.Integer('monto neto sin iva',
          )
-    montoIvaRecuperable = fields.Double(
+    montoIvaRecuperable = fields.Float(
         'monto con iva incluido',
         default=0.19,
         readonly=True
